+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Handle optional and rest parameters in function type expressions.
 
 ## [0.3.3] - 2017-12-18
 - Pin Analyzer version for upcoming major refactor.